refactor: hash-to-point
